+++ R/ensemble_fselect.R
@@ -92,7 +92,7 @@ ensemble_fselect = function(
   assert_task(task)
   assert_learners(as_learners(learners), task = task)
   assert_resampling(init_resampling)
-  assert_choice(class(init_resampling)[1], choices = c("ResamplingBootstrap", "ResamplingSubsampling"))
+  assert_multi_class(init_resampling, c("ResamplingBootstrap", "ResamplingSubsampling"))
   assert_resampling(inner_resampling)
   assert_measure(inner_measure, task = task)
   assert_measure(measure, task = task)
@@ -162,7 +162,7 @@ ensemble_fselect = function(
   set(scores, j = "uhash", value = NULL)
 
   # extract importance scores if RFE optimization was used
-  if (class(fselector)[1] == "FSelectorBatchRFE") {
+  if (inherits(fselector, "FSelectorBatchRFE")) {
     imp_scores = map(afss, function(afs) {
       afs$fselect_result$importance[[1]]
     })