+++ tests/testthat/test_ensemble_fselect.R
@@ -423,3 +423,35 @@ test_that("different callbacks can be set", {
   expect_true(all(efsr$benchmark_result$score()$learner[[1]]$fselect_instance$archive$data$callback_active))
   expect_null(efsr$benchmark_result$score()$learner[[3]]$fselect_instance$archive$data$callback_active)
 })
+
+test_that("efs accepts subclasses of the supported resamplings and fselectors", {
+  ResamplingSubsamplingCustom = R6::R6Class("ResamplingSubsamplingCustom", inherit = mlr3::ResamplingSubsampling)
+  FSelectorBatchRFECustom = R6::R6Class("FSelectorBatchRFECustom", inherit = FSelectorBatchRFE)
+
+  efsr = ensemble_fselect(
+    fselector = FSelectorBatchRFECustom$new(),
+    task = tsk("sonar"),
+    learners = lrns(c("classif.rpart", "classif.featureless")),
+    init_resampling = ResamplingSubsamplingCustom$new(),
+    inner_resampling = rsmp("cv", folds = 3),
+    inner_measure = msr("classif.ce"),
+    measure = msr("classif.ce"),
+    terminator = trm("none")
+  )
+
+  expect_names(names(efsr$result), must.include = "importance")
+
+  expect_error(
+    ensemble_fselect(
+      fselector = fs("random_search"),
+      task = tsk("sonar"),
+      learners = lrns("classif.rpart"),
+      init_resampling = rsmp("cv", folds = 2),
+      inner_resampling = rsmp("cv", folds = 3),
+      inner_measure = msr("classif.ce"),
+      measure = msr("classif.ce"),
+      terminator = trm("evals", n_evals = 3)
+    ),
+    "init_resampling"
+  )
+})